release pwsh action as non-experimetnal
